requirements.txt
pyttsx3==2.90
pyPdf==1.13
libespeak-dev